Organize workspace-related components
