+++ src/renderer/components/index.js
@@ -1,3 +1,4 @@
+export { default as AnswerDistributionPanel } from './workspace/AnswerDistributionPanel';
 export { default as AppMessage } from './AppMessage';
 export { default as BarChart } from './charts/BarChart';
 export { default as CountsWidget } from './charts/CountsWidget';
@@ -11,15 +12,17 @@ export { default as InterviewWidget } from './charts/InterviewWidget';
 export { default as TimeSeriesChart } from './charts/TimeSeriesChart';
 export { default as Modal } from './Modal';
 export { default as Overflow } from './Overflow';
-export { default as PanelItem } from './PanelItem';
+export { default as PanelItem } from './workspace/PanelItem';
 export { default as PairPrompt } from './pairing/PairPrompt';
 export { default as PairPin } from './pairing/PairPin';
 export { default as PairedDeviceModal } from './PairedDeviceModal';
 export { default as PieChart } from './charts/PieChart';
-export { default as ProtocolPanel } from './ProtocolPanel';
+export { default as ProtocolPanel } from './workspace/ProtocolPanel';
 export { default as Scrollable } from './Scrollable';
 export { default as ScrollingPanelItem } from './ScrollingPanelItem';
-export { default as ServerPanel } from './ServerPanel';
-export { default as SessionHistoryPanel } from './SessionHistoryPanel';
-export { default as SessionPanel } from './SessionPanel';
+export { default as ServerPanel } from './workspace/ServerPanel';
+export { default as SessionHistoryPanel } from './workspace/SessionHistoryPanel';
+export { default as SessionPanel } from './workspace/SessionPanel';
+export { default as SortablePanel } from './workspace/SortablePanel';
+export { default as SortablePanels } from './workspace/SortablePanels';
 export { default as TabBar } from './TabBar';
