+++ src/renderer/components/workspace/PanelItem.js
