+++ src/renderer/components/workspace/SortablePanel.js
