+++ src/renderer/components/workspace/SortablePanels.js
