+++ src/renderer/components/workspace/__tests__/AnswerDistributionPanel-test.js
