+++ src/renderer/components/workspace/__tests__/PanelItem-test.js
