+++ src/renderer/components/workspace/__tests__/ProtocolPanel-test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 import React from 'react';
 import { shallow } from 'enzyme';
-import { mockProtocol } from '../../../../config/jest/setupTestEnv';
+import { mockProtocol } from '../../../../../config/jest/setupTestEnv';
 
 import ProtocolPanel from '../ProtocolPanel';
