+++ src/renderer/components/workspace/__tests__/SessionHistoryPanel-test.js
