+++ src/renderer/components/workspace/__tests__/SessionPanel-test.js
