+++ src/renderer/components/workspace/__tests__/SortablePanel-test.js
