+++ src/renderer/components/workspace/__tests__/SortablePanels-test.js
