+++ src/renderer/containers/index.js
@@ -1,8 +1,6 @@
 export { default as App } from './App';
 export { default as ExportScreen } from './ExportScreen';
-export { default as InterviewStatsPanel } from './InterviewStatsPanel';
 export { default as OverviewScreen } from './OverviewScreen';
 export { default as PairDevice } from './PairDevice';
-export { default as ProtocolCountsPanel } from './ProtocolCountsPanel';
 export { default as SettingsScreen } from './SettingsScreen';
-export { default as WorkspaceScreen } from './WorkspaceScreen';
+export { default as WorkspaceScreen } from './workspace/WorkspaceScreen';
