+++ src/renderer/containers/workspace/InterviewStatsPanel.js
@@ -1,9 +1,9 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 
-import InterviewWidget from '../components/charts/InterviewWidget';
-import withApiClient from '../components/withApiClient';
-import { formatDecimal } from '../utils/formatters';
+import InterviewWidget from '../../components/charts/InterviewWidget';
+import withApiClient from '../../components/withApiClient';
+import { formatDecimal } from '../../utils/formatters';
 
 const shapeStatsData = ({ nodes = {}, edges = {} }) => ([
   { name: 'Node count',
